requirements.txt
gradio~=3.50.0
openai~=0.28.1
tiktoken~=0.5.1
pypdf2~=3.0.1
python-dotenv~=1.0.0
weaviate-client~=3.24.2
flask~=3.0.0
waitress~=2.1.2
uvicorn~=0.23.2
fastapi~=0.104.0
requests~=2.31.0
oauth2client
werkzeug~=3.0.0
SQLAlchemy~=2.0.22